# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CursorPaginatedResponse_TrackSampleResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TrackSampleResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "CursorPaginatedResponse[TrackSampleResponse]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TrackSampleBatchRequest": {
        "properties": {
          "samples": {
            "items": {
              "$ref": "#/components/schemas/TrackSampleRequest"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Samples",
            "type": "array"
          }
        },
        "required": [
          "samples"
        ],
        "title": "TrackSampleBatchRequest",
        "type": "object"
      },
      "TrackSampleBatchResponse": {
        "description": "The four buckets add up to the request length.\n\nThat lets a client retire a whole queue slice and still reconcile its local\npoint count against the server's.",
        "properties": {
          "accepted_samples": {
            "title": "Accepted Samples",
            "type": "integer"
          },
          "discarded_samples": {
            "title": "Discarded Samples",
            "type": "integer"
          },
          "duplicate_samples": {
            "title": "Duplicate Samples",
            "type": "integer"
          },
          "filtered_samples": {
            "title": "Filtered Samples",
            "type": "integer"
          }
        },
        "required": [
          "accepted_samples",
          "filtered_samples",
          "duplicate_samples",
          "discarded_samples"
        ],
        "title": "TrackSampleBatchResponse",
        "type": "object"
      },
      "TrackSampleDeleteRequest": {
        "properties": {
          "sample_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 5000,
            "minItems": 1,
            "title": "Sample Ids",
            "type": "array"
          }
        },
        "required": [
          "sample_ids"
        ],
        "title": "TrackSampleDeleteRequest",
        "type": "object"
      },
      "TrackSampleDeleteResponse": {
        "properties": {
          "deleted_count": {
            "title": "Deleted Count",
            "type": "integer"
          }
        },
        "required": [
          "deleted_count"
        ],
        "title": "TrackSampleDeleteResponse",
        "type": "object"
      },
      "TrackSampleModeUpdateRequest": {
        "properties": {
          "sample_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 5000,
            "minItems": 1,
            "title": "Sample Ids",
            "type": "array"
          },
          "travel_mode": {
            "$ref": "#/components/schemas/TravelMode"
          }
        },
        "required": [
          "sample_ids",
          "travel_mode"
        ],
        "title": "TrackSampleModeUpdateRequest",
        "type": "object"
      },
      "TrackSampleModeUpdateResponse": {
        "properties": {
          "updated_count": {
            "title": "Updated Count",
            "type": "integer"
          }
        },
        "required": [
          "updated_count"
        ],
        "title": "TrackSampleModeUpdateResponse",
        "type": "object"
      },
      "TrackSampleRequest": {
        "properties": {
          "accuracy_meters": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy Meters"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "travel_mode": {
            "$ref": "#/components/schemas/TravelMode",
            "default": "UNKNOWN"
          }
        },
        "required": [
          "id",
          "recorded_at",
          "latitude",
          "longitude"
        ],
        "title": "TrackSampleRequest",
        "type": "object"
      },
      "TrackSampleResponse": {
        "properties": {
          "accuracy_meters": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy Meters"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "travel_mode": {
            "$ref": "#/components/schemas/TravelMode"
          }
        },
        "required": [
          "id",
          "recorded_at",
          "latitude",
          "longitude",
          "accuracy_meters",
          "travel_mode"
        ],
        "title": "TrackSampleResponse",
        "type": "object"
      },
      "TrackingSessionCreateRequest": {
        "properties": {
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "started_at"
        ],
        "title": "TrackingSessionCreateRequest",
        "type": "object"
      },
      "TrackingSessionEndRequest": {
        "properties": {
          "ended_at": {
            "format": "date-time",
            "title": "Ended At",
            "type": "string"
          }
        },
        "required": [
          "ended_at"
        ],
        "title": "TrackingSessionEndRequest",
        "type": "object"
      },
      "TrackingSessionListResponse": {
        "properties": {
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/TrackingSessionResponse"
            },
            "title": "Sessions",
            "type": "array"
          }
        },
        "required": [
          "sessions"
        ],
        "title": "TrackingSessionListResponse",
        "type": "object"
      },
      "TrackingSessionResponse": {
        "properties": {
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "recorded_by_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded By User Id"
          },
          "sample_count": {
            "title": "Sample Count",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "started_at",
          "ended_at",
          "recorded_by_user_id",
          "sample_count"
        ],
        "title": "TrackingSessionResponse",
        "type": "object"
      },
      "TravelMode": {
        "description": "Transport mode shared by itinerary legs, post routes, and GPS samples.\n\nValues are stored as plain strings in ``String(32)`` columns, so adding a\nmember is a code change only and never needs a migration.",
        "enum": [
          "UNKNOWN",
          "WALK",
          "BIKE",
          "MOTORCYCLE",
          "CAR",
          "BUS",
          "TRAIN",
          "FERRY",
          "FLIGHT",
          "OTHER"
        ],
        "title": "TravelMode",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "description": "Validation error location path component."
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "OpenVoyage API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/trips/{trip_id}/tracking/samples/delete": {
      "post": {
        "operationId": "delete_track_samples_api_v1_trips__trip_id__tracking_samples_delete_post",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackSampleDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackSampleDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Track Samples",
        "tags": [
          "tracking"
        ]
      }
    },
    "/api/v1/trips/{trip_id}/tracking/samples/travel-mode": {
      "patch": {
        "operationId": "update_track_sample_modes_api_v1_trips__trip_id__tracking_samples_travel_mode_patch",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackSampleModeUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackSampleModeUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Track Sample Modes",
        "tags": [
          "tracking"
        ]
      }
    },
    "/api/v1/trips/{trip_id}/tracking/sessions": {
      "get": {
        "operationId": "list_tracking_sessions_api_v1_trips__trip_id__tracking_sessions_get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingSessionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Tracking Sessions",
        "tags": [
          "tracking"
        ]
      }
    },
    "/api/v1/trips/{trip_id}/tracking/sessions/{session_id}": {
      "delete": {
        "operationId": "delete_tracking_session_api_v1_trips__trip_id__tracking_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Tracking Session",
        "tags": [
          "tracking"
        ]
      },
      "patch": {
        "operationId": "end_tracking_session_api_v1_trips__trip_id__tracking_sessions__session_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackingSessionEndRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "End Tracking Session",
        "tags": [
          "tracking"
        ]
      },
      "post": {
        "operationId": "create_tracking_session_api_v1_trips__trip_id__tracking_sessions__session_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackingSessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Tracking Session",
        "tags": [
          "tracking"
        ]
      }
    },
    "/api/v1/trips/{trip_id}/tracking/sessions/{session_id}/samples": {
      "get": {
        "operationId": "list_track_samples_api_v1_trips__trip_id__tracking_sessions__session_id__samples_get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 2000,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPaginatedResponse_TrackSampleResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Track Samples",
        "tags": [
          "tracking"
        ]
      }
    },
    "/api/v1/trips/{trip_id}/tracking/sessions/{session_id}/samples/batch": {
      "post": {
        "operationId": "upload_track_samples_api_v1_trips__trip_id__tracking_sessions__session_id__samples_batch_post",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackSampleBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackSampleBatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Upload Track Samples",
        "tags": [
          "tracking"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}